+++ src/main/java/com/roadtocda/twiplon/model/LikesId.java
@@ -4,23 +4,23 @@ import java.io.Serializable;
 import java.util.Objects;
 
 public class LikesId implements Serializable {
-	private static final long serialVersionUID = 1L;
+	private static final Long serialVersionUID = 1L;
 	
-	private Long id_user;
-    private Long idpost;
+	private int id_user;
+    private int idpost;
     
     // constructeur, getters et setters (à générer ou écrire manuellement)
     
-	public Long getId_user() {
+	public int getId_user() {
 		return id_user;
 	}
-	public void setId_user(Long id_user) {
+	public void setId_user(int id_user) {
 		this.id_user = id_user;
 	}
-	public Long getIdpost() {
+	public int getIdpost() {
 		return idpost;
 	}
-	public void setIdpost(Long idpost) {
+	public void setIdpost(int idpost) {
 		this.idpost = idpost;
 	} 
 	
@@ -42,5 +42,4 @@ public class LikesId implements Serializable {
 		return Objects.equals(id_user, other.id_user) && Objects.equals(idpost, other.idpost);
 	}
     
-	
 }
